requirements.txt
editdistance==0.4
alphabet-detector
nltk
py-espeak-ng